+++ src/components/elements/Tag/index.tsx
@@ -1,4 +1,4 @@
-import { BrandColor, FontSize } from "@/type";
+import { BrandColor, FontSize } from "@/types";
 
 interface TagProps {
   text: string;
