+++ src/components/module/CompanyInfo/CompanyInfoForm.tsx
@@ -8,7 +8,7 @@ import {
   FormItem,
   FormLabel,
 } from "@/components/ui/form";
-import { CompanyDetailResponse, Success } from "@/type";
+import { CompanyDetailResponse, Success } from "@/types";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useState } from "react";
 import { useForm } from "react-hook-form";
